requirements.txt
pandas>=2.0.0
polars>=0.20.0
pyarrow>=14.0.0
pyyaml>=6.0
